refactor(add-rework): move exercise add files to exercise_add directory
re #105
